Change max_property_value value string
Some higher-level language EditorConfig libraries with multiple return
types might interpret the string "0000...1" as the type "number" and
value "1" (e.g: Lua).

When printed the result will be "1" and not "000...1" and the test fails.

Because the type of the return value is not the purpose of this test,
just the string length, change the string used to a value that should
never be interpreted as a number and is still suitable for the
max_property_value test.